+++ bin/bin-src/cli.js
@@ -1,0 +1,2 @@
+#!/usr/bin/env node
+require('ack-angular-fx/bin/real-cli');